Consolidate some setup_app functions, fix some tests
